+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * ADD MISSION Intercept Transport from Tbilisi to Gudauta
+* ADD MISSION Intercept RECON Patrol from Tbilisi to Gudauta (on demand / scalable)
 * ADD SITUATION SOCHI RED airport
 * ADD SITUATION GUDAUTA BLUE airport
 * FIX RED-AI Kutaisi-F86-ia-1 AirToAir rules